Create Move Zeroes_leetcode_283.cpp
I used the two-pointer concept: one pointer to track the position of non-zero elements (f), and the other to iterate through the current elements (s). Whenever a non-zero element is found, it is swapped with the element at the position indicated by the first pointer, and then the first pointer is incremented.