Create Find All Duplicates in an Array_leetcode_442.cpp
We can take advantage of the fact that elements in the nums array are in the range from 1 to n. By traversing the array, we can set the value at the index corresponding to each element's absolute value to negative. Specifically, if the value of an element is n, its corresponding index will be n-1. If we encounter a position that is already marked as negative, it indicates that the number is a duplicate.