Create Merge Sorted Array_leetcode_88.cpp
We can approach this in two ways. The first is by using an extra array to store the sorted elements from both arrays, and then transferring the sorted elements back to the first array. 
The second approach is in-place sorting, which involves comparing elements from the end of both arrays and placing the greatest values at the end of the first array.